add  添加视频点赞和评论

+++ huomai-business/src/main/java/com/huomai/business/domain/HuomaiVideo.java
@@ -77,20 +77,22 @@ public class HuomaiVideo implements Serializable {
 	/**
 	 * 点赞量
 	 */
-	private Long starNum;
+	private int starNum;
 
 	/**
 	 * 评论量
 	 */
-	private Long commentNum;
+	private int commentNum;
 
 	/**
 	 * 播放量
 	 */
-	private Long viewNum;
+	private int viewNum;
 
-	//分享量
-	private Long shareNum;
+	/**
+	 * 分享量
+	 */
+	private int shareNum;
 
 	/**
 	 * 关键词ID (按","逗号拼接)
